リファクタリング (#32)
